% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_leg_pos.R
\name{best.leg.pos}
\alias{best.leg.pos}
\title{Find best legend position in a plot}
\usage{
best.leg.pos(x, y)
}
\arguments{
\item{x}{The x data or the xy dataset}

\item{y}{Optional. The y data}
}
\description{
Quick function to automatically add a legend in the plot quarter which contains the less points
}
\keyword{internal}
