% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize.R
\name{binarize}
\alias{binarize}
\title{Binarize a vector}
\usage{
binarize(x, zero = NA, one = "else", drop)
}
\arguments{
\item{x}{The vector to binarize}

\item{zero}{The pattern of zeros to be find in x. By default, set to NA. Can be anything.}

\item{one}{The pattern of ones to be find in \code{x}. By default, set to \code{"else"}, which means all that is not of the \code{zero} pattern. Otherwise, can be anything.}

\item{drop}{If there are values not fitting both previous patterns, theses values are returned to NA. By default (\code{drop=TRUE}), these NA are left in the output vector. If you want to discard them, please specify \code{drop=TRUE}.}
}
\value{
The binarized vector of the input vector, with only 0 and 1 (and NA if applicable).
}
\description{
This function transforms a vector to a vector composed of zeros and ones.
}
\details{
For given zero and one 'templates', the function transforms each one in zero or in one and return a 0/1 vector.
}
\examples{

set.seed(1)
x<-runif(100,0,1) # Create a vector of random values
x[as.integer(runif(10,1,100))]<-NA # Randomly replace some of these by NA's
binarize(x) # Binarize it with default option
all(which(is.na(x))==which(binarize(x)==0)) # Check if all NA are zero
all(which(!is.na(x))==which(binarize(x)==1)) # Check if all non-NA are one

x<-c(rep(0,20),2,rep(1,20)) # Create a vector with three repeated numbers: 0, 1, and 2
binarize(x,zero=0,one=1) # Binarize it by saying that zeros are 0, and ones are 1 (thus, two is not allocated, and it is replaced by NA)
binarize(x,zero=0,one=1,drop=TRUE) # Dropping the NA value (which are the no 0 or 1)
length(x) # How many numbers are in x
length(binarize(x,zero=0,one=1)) # How many numbers if we binarize with default (normally, same length than x, because the 2 is replaced by NA)
length(binarize(x,zero=0,one=1,drop=FALSE)) # If we decide do don't drop, the NA is still there, but there is no warnings
length(binarize(x,zero=0,one=1,drop=TRUE)) # If we decide to drop, the NA is removed, and the length is shorter of one

}
