% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxgroups.R
\name{boxgroups}
\alias{boxgroups}
\title{Multivariate and multi-subgroup boxplots}
\usage{
boxgroups(
  data,
  groups,
  box.opt = NULL,
  points.opt = NULL,
  box.width = 0.3,
  group.plot = c("min", "max"),
  group.plot.type = c("b", "l"),
  names = NA,
  ticks = FALSE,
  x.gap = c(0, 0),
  prop = FALSE,
  range = NA
)
}
\arguments{
\item{data}{A matrix or data frame to input.}

\item{groups}{A vector of same length than the number of rows of \code{data} to distinguish subgroups for each variable.}

\item{box.opt}{Optional. Plotting options to be passed for the \code{\link[graphics]{boxplot}} function. Must be a list, the names being those of the parameter of this function.}

\item{points.opt}{Optional. Plotting options to be passed for the \code{\link[graphics]{points}} function. Must be a list, the names being those of the parameter of this function.}

\item{box.width}{The relative width of the boxes, which also governs the gap between the box and the min-max lines of the variable.}

\item{group.plot}{The type of points to be drawn to represent the variation of each subgroup in each variable. By default set to the minimal and maximal range of the subgroup to show its entire range.}

\item{group.plot.type}{The type of plot to be drawn for the plot of each group alongside each boxplot. See \code{plot} argument of the \code{\link[graphics]{plot.default}} function.}

\item{names}{The names to be displayed under each boxplot + lines. Can be the names of the columns of \code{data}. If nothing is provided, will be alphabetical letters.}

\item{ticks}{Logical. Should the ticks be displayed above each name? By default, \code{ticks=FALSE}.}

\item{x.gap}{Gap between the plotted data and the left and right axes. Default set to \code{c(0,0)}; if a single number is provided, automatically meant as left gap only.}

\item{prop}{Logical. Do the boxplots need to be proportional with each other? By default, natural values are plotted (\code{prop=FALSE}).}

\item{range}{If \code{prop=TRUE}, the range of values for rescaling the variables. By default set to 0-1.}
}
\description{
This function plots boxplots for various variables associated with ranges for various subgroups found into each variable
}
\details{
This function first plot the boxplots of the different variables inputted. Then, bars of each of the defined subgroups in each variable are plotted alongside each boxplot, showing desired values describing each subgroup variation within each variable (minimal and maximal values by default, but there can also be other ones detailed below).
Graphical options to the \code{boxplot} and \code{points} formulas can be passed to customize the plot.
There is also a way to rescale all variables in order to be able to compare the variables and the distribution of each subgroup.
\cr
There are various ways to describe the variation of each subgroup alongside the global variable boxplot, which are not exclusive. One needs to specify them in the \code{group.plot} argument with a character vector which can contain:
\describe{
 \item{\code{"min"} and \code{"max"}}{the extreme range, encompassing all points}
 \item{\code{"mean"}}{the average value of the whole subgroup}
 \item{\code{"median"}}{the median of the whole subgroup}
 \item{\code{"sd-X"}}{to obtain points from both sides of the mean by a proportion \code{X} (i.e., a number) of the standard deviation of the whole subgroup}
 \item{\code{"q-X"}}{to obtain the \code{X}th percent quantile (i.e., a percentage)}
 \item{\code{"all"}}{to show all points of the subgroup}
}
}
\examples{
A<-runif(30,1,10)
B<-rexp(30)
C<-rnorm(30,0,1)
data<-data.frame(A=A,B=B,C=C)
groups<-sample(letters[1:4],30,replace=TRUE)
boxgroups(data=data,groups=groups,group.plot = c("all"))
boxgroups(data=data,groups=groups,points.opt=list("pch"=21,"col"=c("blue","red","green3","orange"),"bg"=c("blue","red","green3","orange"),"lwd"=2))

}
