% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasting_palette.R
\name{contrasting.palette}
\alias{contrasting.palette}
\title{Function to give a vector of 22 contrasting colors}
\usage{
contrasting.palette(value, black.and.white = TRUE, sequential = TRUE)
}
\arguments{
\item{black.and.white}{Logical. Choose ot include or not black and white and theend of the list of colors. By default set to TRUE.}

\item{sequential}{Logical. If the value is a single number, choose if you want to have all colors between the first and the value, or if you want to have the color at the value. By default set too TRUE.}
}
\description{
This function is simply a vector, as rainbow() is for instance. It returns 20 simple and distinct colors found by Sasha Trubetskoy + black and white.
}
\seealso{
For further details on this set of 20 of the 22 colors, see \url{https://sashat.me/2017/01/11/list-of-20-simple-distinct-colors/}
}
