% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasting_palette.R
\name{contrasting_palette}
\alias{contrasting_palette}
\title{Function to give a vector of 20 contrasting colors}
\usage{
contrasting_palette()
}
\description{
This function is simply a vector, as rainbow() is for instance. It returns 20 simple and distinct colors found by Sasha Trubetskoy.
}
\seealso{
For further details on this set of 20 colors, see \url{https://sashat.me/2017/01/11/list-of-20-simple-distinct-colors/}
}
