% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{create.tree}
\alias{create.tree}
\title{Create a phylogenetic tree}
\usage{
create.tree(nbtaxa, taxa, age_taxa, nbnodes, age_nodes, format)
}
\arguments{
\item{nbtaxa}{Desired number of taxa; optional, especially if either \emph{taxa} or \emph{age_taxa} parameters are present}

\item{taxa}{A vector specifying the name of \strong{all} desired taxa; if not all taxa names are known, rather do not specify the known names yet and specify them by running the function without this parameter}

\item{age_taxa}{A vector specifying the age of \strong{all} desired taxa; if not all taxa ages are known, rather do not specify the known ages yet and specify them by running the function without this parameter}

\item{nbnodes}{Desired number of nodes; optional, especially if \emph{age_nodes} is present}

\item{age_nodes}{A vector specifying the age of \strong{all} nodes; if not all node ages are known, rather do not specify the known ages yet and specify them by running the function without this parameter}

\item{format}{The format of the output; can be an object of class \emph{phylo} by specifying "phylo" or "phylo object" or simply a newick/parenthetic text by specifying "newick", "NEWICK" or "parenthetic"}
}
\value{
Returns either an object of class \emph{phylo} or a text in newick/parenthetic format. If the latter, the tree can also be saved during performing the function.
}
\description{
This function creates a tree, more a less stepwise, depending on the data in the function.
The function can be run without arguments.
}
