% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_factor.R
\name{extract.from.factor}
\alias{extract.from.factor}
\title{Extract a factor vector with some levels of an initial factor vector}
\usage{
extract.from.factor(factor_vector, factor_levels, x = NA, output)
}
\arguments{
\item{f1}{The initial factor vector}

\item{f2}{The vector containing the levels to extract}
}
\value{
A factor vector with only the specified levels in f2
}
\description{
This function permits to easily have a reduced factor vector containing some of the original levels
}
\examples{
factor_vector<-as.factor(c(rep("A",10),rep("B",5),rep("A",4),rep("C",6)))
factor_levels<-c("B","C")
x<-c(rep(10,10),rep(5,5),rep(11,4),rep(9,6))

extract.from.factor(factor_vector,factor_levels,output="factor")
extract.from.factor(factor_vector,factor_levels,output="logical")
extract.from.factor(factor_vector,factor_levels,x,output="data")

}
