% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_cols.R
\name{freq.cols}
\alias{freq.cols}
\title{Get color palette depending on frequencies within data of given set of colors}
\usage{
freq.cols(cols, ncols, freqs, lims, type, values)
}
\arguments{
\item{cols}{The vector of colors to be resized given data and frequencies}

\item{ncols}{The number of desired colors for the new vector of colors with adjusted color proportions}

\item{freqs}{The inputted frequencies of the colors. Can be a character (\code{"equal"} or \code{"even"}, but actually any character, giving equal frequencies to all colors) or a numeric (specifying proportions or bounds to use)}

\item{lims}{The limits to use to compute bounds (and then adjusted frequencies) for each color. Set by default to the range of \code{values}}

\item{type}{If specifying proportions to \code{freqs}, whether they represent fractions of the width between \code{lims} (\code{type="width"}) or real proportions of the population of \code{values} (\code{type="proportion"})}

\item{values}{If \code{type="proportion"}, the data to use to get the real population proportions and then the adjusted color frequencies}
}
\description{
This function uses a given set of colors and frequencies of them within given dataset and range to derive a set of a given number of colors fitting with the inputted frequencies.
}
\details{
This function mainly relies on \link[ULT]{discrete.palette} to resize the inputted colors taking into account their frequencies.
The frequencies \code{freqs} can be given explicitly (if specifying numeric values), otherwise they are considered to be all equal (if specifying a character).
Numeric frequencies can be either proportions or bounds between \code{lims}. In both cases, they depend on a given set of \code{values} and on limits \code{lims} to consider (by default the range of the dataset).
If frequencies are proportions, they are converted to bounds to determine the effective width between them (or to one of the limits) relative to the inter-limits width. Hence, bounds can also be directly provided, if known.
Such relative widths then serve as the \code{freqs} argument of \link[ULT]{discrete.palette}.
If proportions are provided (explicit numeric or just character), it is also possible to say with \code{type} whether these proportions are between-limits width proportion or \code{values} population proportion. In the latter case, quantiles corresponding to the given frequencies \code{freqs} are hence used as bounds.
The output will use adjusted frequencies (given the inputted one and the width/proportion choice if applicable) to resize the inputed colors \code{cols} as a new vector of colors of size \code{ncols}
}
