% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClade.R
\name{getClade}
\alias{getClade}
\title{Get the node of a given monophyletic group.}
\usage{
getClade(tree, taxa, output = "number")
}
\arguments{
\item{tree}{The phylogenetic tree to provide}

\item{taxa}{The taxa (tips only) forming the monophyletic group (i.e., a monophyletic group is at least composed of two tips).}

\item{output}{Optional. The format of the output (i.e., the node). Can be either the node number according to the \code{nodelabels} function (\code{output="number"}, the default), or its number across nodes (\code{output="node number"}), or its name if this has been specified in the tree (\code{output="name"}).}
}
\value{
The number or name of the node containing the given tips.
}
\description{
This function aims to output the monophyletic group (node number or name) that contains a given set of terminal taxa, and only them.
}
\details{
This function heavy relies on the function \code{getDescendants} from the package phytools, which give all phylogenetical units (i.e., tips and nodes) that are contained in a given clade.
}
\examples{
newick_tree<-c("((((A,B),C),(D,E)),F);") # Create a simple tree in NEWICk format
tree<-ape::read.tree(text = newick_tree) # Read it
plot(tree)
ape::nodelabels() # Plot it together with node numbers

taxa<-c("A","B","C") # A monophyletic group
getClade(tree,taxa) # The actual 9 on the phylogeny
getClade(tree,taxa,output="node number") # The node number three going from the root

tree2<-ape::makeNodeLabel(tree,method="number",prefix="N") # Adding names to the nodes
plot(tree2)
ape::nodelabels(tree2$node.label)
getClade(tree2,taxa,output="name")

}
