% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphospace.R
\name{morphospace}
\alias{morphospace}
\title{Function to create a "morphospace"}
\usage{
morphospace(data, plot.function, output = "plot", ...)
}
\arguments{
\item{data}{Data frame with two vectors}

\item{plot.function}{Type of function used to plot the polygon: "points" or "polygon". Default is "polygon".}

\item{...}{graphical arguments, depend of the plot.function choosed}
}
\description{
This functions performs a "morphospace" of a dataset of two vectors: it connects extremal points to plot the largest polygon possible. This is a rather generic function and can be applied on datasets no matter of which are the vectors. The dataset just needs two vectors: firstly, plot the two vectors against each other (second vector in y-axis; first vector in x-axis), then run morphospace to superimpose a polygon to the plot.
}
\details{
The function firstly identify the most extreme point on the x-axis, which is the starting point of the polygon.
Secondly, the function finds the second point of the polygon which is defined as having the steepest slope with the first point among all other points.
Thirdly, the function iteratively calculates the angle (using the same code than the function Angle in the package LearnGeom) between the slopes preceeding and succeeding the i-point (starting from the second defined point): the slope preceeding the i-point is unique (it is defined by the actual i point and the preceeding i-1 point), each slope succeeding the i-point is considered (with all other points). The largest angle is retained together with the associated i+1 point.
At the end, the function creates another data.frame of two vectors which each point surrounding the whole scatterplot; the function creates the polygon based on these points.
}
\examples{

vec1_test<-sample.int(15,replace=TRUE)
vec2_test<-sample.int(15,replace=TRUE)
df_test<-data.frame(vec1_test,vec2_test)
plot(df_test)
morphospace(df_test)

}
