% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multkw_perm.R
\name{multkw.perm}
\alias{multkw.perm}
\title{Extended Multivariate Kruskal-Wallis test with missing data}
\usage{
multkw.perm(nmc, group, y, r, weight)
}
\arguments{
\item{nmc}{Number of Monte-Carlo permutations to do.}

\item{group}{The factorial variable that subsets the dataset in groups. Can be a character vector, a factorial vector or an integer/numeric vector.}

\item{y}{The dataset of n numeric(or integer) variables.}

\item{r}{Optional. The missing data pattern to be applied. If dataset has NA and if the missing data pattern is the distribution of the NA's in the dataset, r is optional and is automatically computed.}

\item{weight}{Optional. The weighting scheme to be used to compute the final value of the test statistic. As test statistics are calculated for each pattern of missingness, there are as statistics as patterns. The final test statistic can thus be the arithmetic mean of each statistic (weight="equal") or the ponderated mean of each statistic relative to the proportion of each missing pattern (weight="prop").}
}
\value{
Returns a list of results of the various multivariate Kruskal-Wallis tests that have been computed.
The results are the test statistics (W2), the degrees of freedom (df) and the p-value of the test statistic.
These three results are given for (1) a "classical" multivariate Kruskal-Wallis test, i.e. on data without missing values; each test statistic is thus followed by a .c for "complete" and (2) a global multivariate Kruskal-Wallis test that takes into account missing values (see details); each test statistic is thus followed by a .m for "missing".
}
\description{
This function computes an "extended" multivariate Kruskal-Wallis test for n numeric variables (which can contain NA's) relative to one factorial variable (that subsets the dataset in groups)
}
\details{
"Likelihood-based" and "permutation-based" multivariate Kruskal-Wallis tests are computed: in large samples, the distribution of the test statistic approximates that of the khi?, but in smaller samples, a more accurate p-value is obtained by computing an "empirical" distribution of the test statistic by doing a Monte-Carlo sampling with permutations.
Firstly, the "multivariate Kruskal-Wallis test with missing data" is computed and are the first half of the outputs; they are the "likelihood-based" test results (see documentation of multkw_m for more details).
Thus, a Monte-Carlo sampling with permutations (by randomly assigning individuals to groups) is computed, and the second half of the outputs are the proportions of results that exceeds the previously observed results (with the likelihood-based test).
}
\examples{
data(airquality)
datamkw<-airquality[,1:4]
multkw(y=datamkw,airquality$Month)
multkw.m(y=datamkw,airquality$Month)
multkw.perm(y=datamkw,airquality$Month,nmc=100)
multkw.perm(y=datamkw,airquality$Month,nmc=10000)

}
\references{
\insertRef{He.etal.2017}{ULT}
}
\seealso{
See chapter 2.2.2 and 4.2 of the \href{http://d-scholarship.pitt.edu/19411/1/Fanyin_ETD_draft_08-06-2013.pdf}{PhD manuscript of Fanyin He} and 'Methodology' of \insertCite{He.etal.2017;textual}{ULT} for more details.
}
