% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multkw.R
\name{multkw}
\alias{multkw}
\title{Multivariate Kruskal-Wallis test}
\usage{
multkw(group, y)
}
\arguments{
\item{group}{The factorial variable that subsets the dataset in groups. Can be a character vector, a factorial vector or an integer/numeric vector.}

\item{y}{The dataset of n numeric(or integer) variables.}
}
\value{
Output is either a list (with "simplify=FALSE") or a vector (with "simplify=TRUE") containing the results of the multivariate Kruskal-Wallis test.
}
\description{
This function computes a multivariate Kruskal-Wallis test for n numeric variables relative to one factorial variable (that subsets the dataset in groups)
}
\details{
A "standard" multivariate Kruskal-Wallis test is computed, deleting all missing data.
}
\examples{
data(airquality)
datamkw<-airquality[,1:4]
multkw(y=datamkw,airquality$Month)

}
\seealso{
See chapter 2.2.1 of the \href{http://d-scholarship.pitt.edu/19411/1/Fanyin_ETD_draft_08-06-2013.pdf}{PhD manuscript of Fanyin He} for more details.
}
