% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_t_test.R
\name{param.t.test}
\alias{param.t.test}
\title{T-test using describing parameters of data}
\usage{
param.t.test(
  mean_x,
  var_x,
  n_x,
  mean_y = NA,
  var_y = NA,
  n_y = NA,
  mu = 0,
  paired = FALSE,
  var.equal = FALSE,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  name_x = "x",
  name_y = "y"
)
}
\value{
As for \code{\link[stats]{t.test}}, a list with class \code{"htest"} containing the following components:
\describe{
\item{\code{statistic}}{the value of the t-statistic}
\item{\code{parameter}}{the degrees of freedom of the t-statistic}
\item{\code{p.value}}{the p-value for the test}
\item{\code{conf.int}}{a confidence interval for the mean appropriate to the specified alternative hypothesis}
\item{\code{estimate}}{the (inputted) estimated mean or means depending on whether it was a one-sample or a two-sample test}
\item{\code{null.value}}{the specified hypothesized value of the mean or mean difference depending on whether it was a one-sample or a two-sample test}
\item{\code{stderr}}{the standard error of the mean/difference, used as denominator in the t-statistic formula}
\item{\code{alternative}}{a character string describing the alternative hypothesis}
\item{\code{method}}{a character string indicating what type of t-test was performed}
\item{\code{data.name}}{a character string giving the (inputted) name(s) of the data}
}
}
\description{
Performs one and two sample t-tests using describing parameters (i.e., mean, variance, length) of data
}
\details{
\code{alternative = "greater"} is the alternative that \code{x} has a larger mean than \code{y}. For the one-sample case: that the mean is positive.
If \code{paired} is \code{TRUE}, an error is returned because a paired t-test can only be done using original data, and the user should perform a t-test using the \code{\link[stats]{t.test}} function
}
\examples{
## Define data
set.seed(1)
A<-runif(10,0,100)
B<-runif(10,50,150)

mean_x<-mean(A)
var_x<-var(A)
n_x<-length(A)
mean_y<-mean(B)
var_y<-var(B)
n_y<-length(B)

## One sample t-test
t.test(A)
param.t.test(mean_x=mean_x,var_x=var_x,n_x=n_x)

## Two sample t-test with equal variances
t.test(A,B,var.equal=TRUE)
param.t.test(mean_x,var_x,n_x,mean_y,var_y,n_y,var.equal=TRUE)

## Two sample t-test with unequal variances
t.test(A,B)
param.t.test(mean_x,var_x,n_x,mean_y,var_y,n_y)
}
\seealso{
\code{\link[stats]{t.test}}
}
