% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_twosided.R
\name{plot.twosided}
\alias{plot.twosided}
\title{Function to plot two y-variables relative to a x-variable}
\usage{
\method{plot}{twosided}(
  x,
  y,
  y2,
  xy.plot.type,
  xy2.plot.type,
  xy.plot.param,
  xy2.plot.param,
  y.axis.param,
  y2.axis.param
)
}
\arguments{
\item{x}{The x-variable to be plotted. Can be either numeric, integer but also factor or character.}

\item{y}{The first y-variable to be plotted (y-axis is plotted to the left). Can be either numeric or integer.}

\item{y2}{The second y-variable to be plotted (y2-axis is plotted to the right). Can also be either numeric or integer.}

\item{xy.plot.type}{The plot function used to plot the x-y relationship: can be "plot" or "boxplot". By default, the formula is plot(x,y).}

\item{xy2.plot.type}{The plot function used to plot the x-y2 relationship: can be "points" or "boxplot". By default, the formula is points(x,y).}

\item{xy.plot.param}{The graphical parameters to use for the x-y plot; must be a list.}

\item{xy2.plot.param}{The graphical parameters to use for the x-y2 plot; must be a list.}

\item{y.axis.param}{The graphical parameters to use for the first/left y axis; must be a list.}

\item{y2.axis.param}{The graphical parameters to use for the second/right y axis; must be a list.}
}
\description{
This function performs a classic plot of a y-variable (y) relative to a x-variable then superimpose a second plot of a second y-variable (y2) relative to the same x-variable
}
\details{
The function basically needs three vectors to be performed: the "x", "y" and "y2" variables.
It also needs the packages 'scales'.
The x-variable can be of any type of vector. Both y and y2 variables must be numeric or integer.
This function firstly perform a plot of x and y (by default, "plot" is used with x and y successively, not a ~ formula) then performs a plot of x and of a y-scaled y2 variable (by default, "points" is used).
The first y axis is plotted to the left, while the second one is plotted to the right.
The second/right y axis is of same size than the first/left y axis, but of different scale.
Plot types can be modified, as well as graphical parameters of the plots and of the two y-axis.
}
\examples{

# For all three variables being numeric:
x<-sort(runif(100,0,1))
y<-sort(rnorm(100,0,1))
y2<-sort(rexp(100))
plot_twosided(x,y,y2,xy.plot.param=list(type="l",col="red",lwd=2),xy2.plot.param=list(type="l",col="blue",lwd=2),y.axis.param=list(col="red"),y2.axis.param=list(col="blue"))

# For character (= factorial) x:
x<-as.factor(sort(rep(letters[1:10],10)))
y<-rexp(100,0.5)
y2<-sort(rep(runif(10,15,25),10))
plot_twosided(x,y,y2,xy.plot.param=list(col="red",border="blue"),xy2.plot.param=list(type="l",col="green3",lwd=2),y.axis.param=list(col="red"),y2.axis.param=list(col="green3"))

}
