% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_multkw_output_JM.R
\name{print.multkw.output}
\alias{print.multkw.output}
\title{Print function for results of multkw functions (multkw, multkw.m and multkw.perm)}
\usage{
\method{print}{multkw.output}(x)
}
\arguments{
\item{x}{An object of class "multkw.output"}
}
\description{
The multivariate Kruskal-Wallis test functions (multkw, multkw.m and multkw.perm) output an object of multkw.output class.
It can be saved to an object, but if not, result is automatically printed using this function.
If saved, just do print.multkw.output(object) to see its results.
}
