% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs_tree.R
\name{subs.tree}
\alias{subs.tree}
\title{Substitute a taxon of a given phylogenetic tree by another tree}
\usage{
subs.tree(back.tree,fore.tree,where,drop.where=TRUE,poly.where=FALSE,stem.edge.length=NULL,node.age=NULL)
}
\arguments{
\item{back.tree}{The tree on which a taxon is substituted by another tree, named "background" tree because it is the general support/scaffold/backbone.}

\item{fore.tree}{The tree to put in the "background" tree, named "foreground" tree because it is the one to put on the other tree.}

\item{where}{The taxon of \code{back.tree} one wants to replace by the \code{fore.tree}. It can be the number or the name of the taxon (tip or node), but it can also be a vector of some tips (allowing for specifying a node without having to search for it and without having to quote every of its descendants).}

\item{drop.where}{Optional. The user can discard the given taxon where to put the \code{fore.tree} (default condition, set to \code{TRUE}), or it can keep it if the user wants to place the \code{fore.tree} on the edge leading to that taxon (if set to \code{FALSE}).}

\item{poly.where}{Optional. The user can chose whether the \code{fore.tree} has to be on a distinct edge (default condition, set to \code{FALSE}) or if it has to be in a polytomy with the specified taxon if the taxon is a node (if set to \code{TRUE}).}

\item{stem.edge.length}{Optional. The length of the edge leading to the \code{fore.tree} in the new tree. Complementary with the \code{node.age} parameter; please do not specify both.}

\item{node.age}{Optional. The age of the 'root' node of the \code{fore.tree} is the new tree. Must be lower than the age of the previous node (if not, a warning message with the age of that previous node is displayed). Complementary with the \code{stem.edge.length} parameter; please do not specify both.}
}
\value{
A tree of class \code{phylo} with both the \code{back.tree} and the \code{fore.tree} merged together.
}
\description{
The goal of this function is to enhance manipualtion between trees, especially for constructing composite trees, by allowing one to substitute a taxon (tip or subtree) of a tree by another tree.
}
\details{
This function replaces a taxon of a given "background" tree, be it a single tip or a node (i.e., part of the tree), by another "foreground" tree.
The age of the node of the "foreground" tree is not necessarily the same than the age of the taxon to replace in the "background" tree; there are therefore options allowing for specifying its age.
}
\examples{
require(ape)
back.tree<-read.tree(text="((((A:1,B:2):0.5,(C:3,(D:4,E:1):2):5):1,(F:3,(G:2.5,(H:2.5,(I:1.5,J:1):1):1):2):1):4,K:10);")
plot(back.tree)
nodelabels()

fore.tree<-read.tree(text="((x:0.5,y:1.5):5,z:11);")
plot(fore.tree)

where<-c("C","E")

plot(subs.tree(back.tree,fore.tree,where,node.age=8))
plot(subs.tree(back.tree,fore.tree,where,stem.edge.length=0.1))
plot(subs.tree(back.tree,fore.tree,where))

}
