% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_arrows.R
\name{tax.arrows}
\alias{tax.arrows}
\title{Draw arrows next to phylogenetic taxa}
\usage{
tax.arrows(
  phy,
  tax,
  side = "topleft",
  angle = 45,
  length,
  offset,
  arrow.fun = "arrows",
  arrow.pars
)
}
\arguments{
\item{phy}{A phylogenetic tree.}

\item{tax}{A vector with the name or the number of the taxa (tips and/or nodes) for plotting the arrows next to.}

\item{side}{A vector to specify the side for plotting the arrows. Can be one \code{1} or \code{"topleft"} (the default), \code{2} or \code{"topright"}, \code{3} or \code{"bottomleft"}, or \code{4} or \code{"bottomright"}.}

\item{angle}{The angle of the arrow in degrees. Can be a value between 0 and 360, somewhat overriding the \code{side} parameter. Default set to 45.}

\item{length}{The length of the arrow. Default is to one tenth of the x span, but the user may try several values since this length if then expressed using both x and y.}

\item{offset}{The length of the distance between the taxon and the end of the arrow. Default is to one tenth of the length, but the user may try several values since this length if then expressed using both x and y.}

\item{arrow.fun}{The function to be used to draw the arrows, keeping in mind that the system used is here by providing x0/x1/y0/y1 coordinates. Default to \code{arrows}, but the user may want to use the [shape::Arrows] function.}

\item{arrow.pars}{A list of arguments to be passed to the arrow function used. See [base::arrows] for the base function arguments; see [shape::Arrows] for the function from the \code{shape} package.}
}
\description{
By giving taxon numbers of a phylogenetic tree and graphical parameters, the function plots the phylogeny (if not already plotted) and arrows next to taxa for graphical purposes.
}
\examples{
require(ape)
set.seed(1)
phy<-rtree(20)
plot(phy)
tax<-c(2,7,32)
tax.arrows(phy,tax)
tax.arrows(phy,tax,angle=0)
tax.arrows(phy,tax,angle=90)
tax.arrows(phy,tax,angle=45,arrow.pars=list(length=0.1,angle=30))
require(shape)
tax.arrows(phy,tax,angle=45,arrow.pars=list(arr.length=0.5,arr.type="triangle"),arrow.fun="Arrows")
tax.arrows(phy,tax,angle=45,arrow.pars=list(arr.length=0.2,arr.type="triangle",arr.col="red"),arrow.fun="Arrows")

}
