% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasting_palette.R
\name{contrasting_palette}
\alias{contrasting_palette}
\title{Function to give a vector of 22 contrasting colors}
\usage{
contrasting_palette(value, black.and.white = TRUE)
}
\arguments{
\item{black.and.white}{Logical. Choose ot include or not black and white and theend of the list of colors. By default set to TRUE.}
}
\description{
This function is simply a vector, as rainbow() is for instance. It returns 20 simple and distinct colors found by Sasha Trubetskoy + black and white.
}
\seealso{
For further details on this set of 20 of the 22 colors, see \url{https://sashat.me/2017/01/11/list-of-20-simple-distinct-colors/}
}
