% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{create.tree}
\alias{create.tree}
\title{Create a phylogenetic tree}
\usage{
create.tree(
  nbtaxa,
  taxa,
  age_taxa,
  nbnodes,
  nodes,
  age_nodes,
  tax_selection,
  ultra,
  format
)
}
\arguments{
\item{nbtaxa}{Desired number of taxa; optional, especially if either \code{taxa} or \code{age_taxa} parameters are present}

\item{taxa}{A vector specifying the name of \strong{all} desired taxa; if not all taxa names are known, rather do not specify the known names yet and specify them by running the function without this parameter}

\item{age_taxa}{A vector specifying the age of \strong{all} desired taxa; if not all taxa ages are known, rather do not specify the known ages yet and specify them by running the function without this parameter}

\item{nbnodes}{Desired number of nodes; optional, especially if \code{age_nodes} is present}

\item{nodes}{A list of vectors; each vector is a node and must contains the names of the "targets", either the taxa (full name) or other nodes (Ni, i being the i-th node). It is advised to name each vector of the list "Ni" (i being the i-th node); if not, nodes are assumed to be hierarchized from oldest to youngest (while by naming nodes, the order does not matter).}

\item{age_nodes}{A vector specifying the age of \strong{all} nodes; if not all node ages are known, rather do not specify the known ages yet and specify them by running the function without this parameter}

\item{tax_selection}{The method to choose taxa if \code{taxa} and/or \code{nodes} are not specified. Can be either \code{BYNAME} (choose taxa by their name) or \code{BYGRAPH} (choose taxa by clicking them)}

\item{ultra}{Logical; if the tree has to be ultrametric (i.e. without specific node ages)}

\item{format}{The format of the output; can be an object of class \code{phylo} by specifying \code{"phylo"} or \code{"phylo object"} or simply a newick/parenthetic text by specifying \code{"newick"}, \code{"NEWICK"} or \code{"parenthetic"}}
}
\value{
Returns either an object of class \code{phylo} or a text in newick/parenthetic format. If the latter, the tree can also be saved during performing the function.
}
\description{
This function creates a tree, more a less stepwise, depending on the data in the function.
The function can be run without arguments.
If the desired phylogeny is already known and prepared, it is advised to prepare and save the elements used in this function, in case of small errors (yet, the function doesn't allow one to recover inputted data if there is an error).
}
