% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_palette.R
\name{discrete.palette}
\alias{discrete.palette}
\title{Function to provide a user-defined discrete color palette}
\usage{
discrete.palette (ncols, cols, prop)
}
\arguments{
\item{ncols}{The number of colors to be outputted.}

\item{cols}{The colors to be used.}

\item{prop}{Optional. The proportions for the repartition of the colors. By default, proportions are all equal.}
}
\value{
A vector that contains \code{ncols} color codes, with all \code{cols} repeated.
}
\description{
This function aims to give a 'discrete' palette, that is, a palette without gradient and with some user-defined repeated colors.
}
\details{
The function takes as input the colors the user wants to repeat, and optionally their proportions if it is not intended to distribute them equally.
The equal distribution may not be fully true: if needed, proportions are rounded to have an integer number of repeats of each color. The goal is rather to output the specified number of colors the user wants.
}
