% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_factor.R
\name{extract.from.factor}
\alias{extract.from.factor}
\title{Extract a factor vector with some levels of an initial factor vector}
\usage{
extract.from.factor(f1, f2)
}
\arguments{
\item{f1}{The initial factor vector}

\item{f2}{The vector containing the levels to extract}
}
\value{
A factor vector with only the specified levels in f2
}
\description{
This function permits to easily have a reduced factor vector containing some of the original levels
}
\examples{
f1<-as.factor(c(rep("A",10),rep("B",5),rep("C",8),rep("A",4)))
f2<-c("A","B")
extract.from.factor(f1,f2)

}
